Add files via upload
Angular2 shifts the mindset from writing a web page application from a single file of code to multiple files, and breaking down the components of a webpage into smaller modules based on their functionality; and eventually bringing all the components together. All of the files work together to synchronize a single web-page.
The components are the small modules that are made up of 4 files(Typescript, HTML, CSS, and Spec.tc), These components are then grouped in a single HTML page.